Revert "refactor: pages structure"
