+++ .env.example
@@ -62,4 +62,3 @@ AWS_BUCKET=
 AWS_USE_PATH_STYLE_ENDPOINT=false
 
 VITE_APP_NAME="${APP_NAME}"
-VITE_API_URL="${APP_URL}/api"
