+++ resources/ts/AppRoutes.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter as Router, useRoutes } from 'react-router-dom';
-import Error from '@/pages/error';
+import Error from '@/pages/Error';
 
 import { useAuth } from '@/hooks/useAuth';
