+++ resources/ts/pages/Admin/Settings/Contracts.tsx
@@ -8,46 +8,62 @@ import { Icon } from '@iconify/react';
 
 import axiosClient from '@/api/axiosClient';
 import { useTranslation } from 'react-i18next';
-import CrudPanel from '@/components/admin/CrudPanel';
+import CrudPanel from '@/components/Admin/CrudPanel';
 
-export default function TaskTypes() {
+export default function Contracts() {
   const [isLoading, setIsLoading] = useState(true);
-  const [taskTypes, setTaskTypes] = useState([]);
+  const [contracts, setContracts] = useState([]);
   const { t } = useTranslation();
 
   useEffect(() => {
-    const fetchTaskTypes = async () => {
+    const fetchContracts = async () => {
       try {
-        const response = await axiosClient.get('/admin/task-types');
-        setTaskTypes(response.data);
+        const response = await axiosClient.get('/admin/contracts');
+        setContracts(response.data);
         setIsLoading(false);
       } catch (error) {
         console.error(error);
       }
     };
 
-    fetchTaskTypes();
+    fetchContracts();
   }, []);
 
   return (
     <>
       <CrudPanel
-        title="admin.pages.taskTypes.title"
-        onCreate={() => console.log('Create new task type')}>
+        title="admin.pages.contracts.title"
+        onCreate={() => console.log('Create new contract')}>
         <DataTable
           loading={isLoading}
-          value={taskTypes}
+          value={contracts}
           paginator
           rows={10}
           stripedRows
           showGridlines>
           <Column
             field="name"
-            header={t('admin.pages.taskTypes.columns.name')}
+            header={t('admin.pages.contracts.columns.name')}
           />
           <Column
-            field="description"
-            header={t('admin.pages.taskTypes.columns.description')}
+            field="start_date"
+            header={t('admin.pages.contracts.columns.start_date')}
+          />
+          <Column
+            field="end_date"
+            header={t('admin.pages.contracts.columns.end_date')}
+          />
+          <Column
+            field="invoice_proposed"
+            header={t('admin.pages.contracts.columns.invoice_proposed')}
+          />
+          <Column
+            field="invoice_agreed"
+            header={t('admin.pages.contracts.columns.invoice_agreed')}
+          />
+          <Column
+            field="invoice_paid"
+            header={t('admin.pages.contracts.columns.invoice_paid')}
           />
           {/* Actions */}
           <Column
