+++ resources/ts/pages/Admin/Settings/ElementTypes.tsx
@@ -8,46 +8,58 @@ import { Icon } from '@iconify/react';
 
 import axiosClient from '@/api/axiosClient';
 import { useTranslation } from 'react-i18next';
-import CrudPanel from '@/components/admin/CrudPanel';
+import CrudPanel from '@/components/Admin/CrudPanel';
 
-export default function ResourceTypes() {
+export default function ElementTypes() {
   const [isLoading, setIsLoading] = useState(true);
-  const [resourceTypes, setResourceTypes] = useState([]);
+  const [elementTypes, setElementTypes] = useState([]);
   const { t } = useTranslation();
 
   useEffect(() => {
-    const fetchResourceTypes = async () => {
+    const fetchElementTypes = async () => {
       try {
-        const response = await axiosClient.get('/admin/resource-types');
-        setResourceTypes(response.data);
+        const response = await axiosClient.get('/admin/element-types');
+        setElementTypes(response.data);
         setIsLoading(false);
       } catch (error) {
         console.error(error);
       }
     };
 
-    fetchResourceTypes();
+    fetchElementTypes();
   }, []);
 
   return (
     <>
       <CrudPanel
-        title="admin.pages.resourceTypes.title"
-        onCreate={() => console.log('Create new resource type')}>
+        title="admin.pages.elementTypes.title"
+        onCreate={() => console.log('Create new element type')}>
         <DataTable
           loading={isLoading}
-          value={resourceTypes}
+          value={elementTypes}
           paginator
           rows={10}
           stripedRows
           showGridlines>
           <Column
             field="name"
-            header={t('admin.pages.resourceTypes.columns.name')}
+            header={t('admin.pages.elementTypes.columns.name')}
+          />
+          <Column
+            field="requires_tree_type"
+            header={t('admin.pages.elementTypes.columns.requires_tree_type')}
           />
           <Column
             field="description"
-            header={t('admin.pages.resourceTypes.columns.description')}
+            header={t('admin.pages.elementTypes.columns.description')}
+          />
+          <Column
+            field="icon"
+            header={t('admin.pages.elementTypes.columns.icon')}
+          />
+          <Column
+            field="color"
+            header={t('admin.pages.elementTypes.columns.color')}
           />
           {/* Actions */}
           <Column
