+++ resources/ts/pages/Admin/Settings/ResourceTypes.tsx
@@ -8,58 +8,46 @@ import { Icon } from '@iconify/react';
 
 import axiosClient from '@/api/axiosClient';
 import { useTranslation } from 'react-i18next';
-import CrudPanel from '@/components/admin/CrudPanel';
+import CrudPanel from '@/components/Admin/CrudPanel';
 
-export default function ElementTypes() {
+export default function ResourceTypes() {
   const [isLoading, setIsLoading] = useState(true);
-  const [elementTypes, setElementTypes] = useState([]);
+  const [resourceTypes, setResourceTypes] = useState([]);
   const { t } = useTranslation();
 
   useEffect(() => {
-    const fetchElementTypes = async () => {
+    const fetchResourceTypes = async () => {
       try {
-        const response = await axiosClient.get('/admin/element-types');
-        setElementTypes(response.data);
+        const response = await axiosClient.get('/admin/resource-types');
+        setResourceTypes(response.data);
         setIsLoading(false);
       } catch (error) {
         console.error(error);
       }
     };
 
-    fetchElementTypes();
+    fetchResourceTypes();
   }, []);
 
   return (
     <>
       <CrudPanel
-        title="admin.pages.elementTypes.title"
-        onCreate={() => console.log('Create new element type')}>
+        title="admin.pages.resourceTypes.title"
+        onCreate={() => console.log('Create new resource type')}>
         <DataTable
           loading={isLoading}
-          value={elementTypes}
+          value={resourceTypes}
           paginator
           rows={10}
           stripedRows
           showGridlines>
           <Column
             field="name"
-            header={t('admin.pages.elementTypes.columns.name')}
-          />
-          <Column
-            field="requires_tree_type"
-            header={t('admin.pages.elementTypes.columns.requires_tree_type')}
+            header={t('admin.pages.resourceTypes.columns.name')}
           />
           <Column
             field="description"
-            header={t('admin.pages.elementTypes.columns.description')}
-          />
-          <Column
-            field="icon"
-            header={t('admin.pages.elementTypes.columns.icon')}
-          />
-          <Column
-            field="color"
-            header={t('admin.pages.elementTypes.columns.color')}
+            header={t('admin.pages.resourceTypes.columns.description')}
           />
           {/* Actions */}
           <Column
