+++ resources/ts/pages/Admin/Settings/Users.tsx
@@ -9,7 +9,7 @@ import { Icon } from '@iconify/react';
 
 import axiosClient from '@/api/axiosClient';
 import { useTranslation } from 'react-i18next';
-import CrudPanel from '@/components/admin/CrudPanel';
+import CrudPanel from '@/components/Admin/CrudPanel';
 
 export default function Users() {
   const [isLoading, setIsLoading] = useState(true);
