+++ resources/ts/pages/Admin/WorkOrders.tsx
@@ -1,53 +1,54 @@
 import { useState, useEffect } from 'react';
-
 import { DataTable } from 'primereact/datatable';
 import { Column } from 'primereact/column';
 import { Button } from 'primereact/button';
-
 import { Icon } from '@iconify/react';
-
 import axiosClient from '@/api/axiosClient';
 import { useTranslation } from 'react-i18next';
-import CrudPanel from '@/components/admin/CrudPanel';
+import CrudPanel from '@/components/Admin/CrudPanel';
 
-export default function ResourceTypes() {
+export default function WorkOrders() {
   const [isLoading, setIsLoading] = useState(true);
-  const [resourceTypes, setResourceTypes] = useState([]);
+  const [workOrders, setWorkOrders] = useState([]);
   const { t } = useTranslation();
 
   useEffect(() => {
-    const fetchResourceTypes = async () => {
+    const fetchWorkOrders = async () => {
       try {
-        const response = await axiosClient.get('/admin/resource-types');
-        setResourceTypes(response.data);
+        const response = await axiosClient.get('/admin/work-orders');
+        setWorkOrders(response.data);
         setIsLoading(false);
       } catch (error) {
         console.error(error);
       }
     };
 
-    fetchResourceTypes();
+    fetchWorkOrders();
   }, []);
 
   return (
     <>
       <CrudPanel
-        title="admin.pages.resourceTypes.title"
-        onCreate={() => console.log('Create new resource type')}>
+        title="admin.pages.workOrders.title"
+        onCreate={() => console.log('Create new work order')}>
         <DataTable
           loading={isLoading}
-          value={resourceTypes}
+          value={workOrders}
           paginator
           rows={10}
           stripedRows
           showGridlines>
           <Column
-            field="name"
-            header={t('admin.pages.resourceTypes.columns.name')}
+            field="date"
+            header={t('admin.pages.workOrders.columns.date')}
           />
           <Column
-            field="description"
-            header={t('admin.pages.resourceTypes.columns.description')}
+            field="status"
+            header={t('admin.pages.workOrders.columns.status')}
+          />
+          <Column
+            field="contract_id"
+            header={t('admin.pages.workOrders.columns.contract_id')}
           />
           {/* Actions */}
           <Column
