+++ resources/ts/pages/Login.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/components/login/LoginForm';
+import LoginForm from '@/components/Login/LoginForm';
 
 export default function Login() {
   return <LoginForm />;
