--- resources/ts/pages/admin/dashboard.tsx
@@ -1,11 +1,0 @@
-import { Welcome } from '@/components/admin/dashboard/Welcome';
-import { Stats } from '@/components/admin/dashboard/Stats';
-
-export default function Dashboard() {
-  return (
-    <>
-      <Welcome />
-      <Stats />
-    </>
-  );
-}
