--- resources/ts/pages/admin/inventory.tsx
@@ -1,3 +1,0 @@
-export default function Inventory() {
-  return null;
-}
