--- resources/ts/pages/admin/workers/index.tsx
@@ -1,3 +1,0 @@
-export default function Workers() {
-  return null;
-}
