--- resources/ts/pages/customer/dashboard.tsx
@@ -1,7 +1,0 @@
-export default function Dashboard() {
-  return (
-    <div>
-      <h1>Customer Dashboard</h1>
-    </div>
-  );
-}
