--- resources/ts/pages/login.tsx
@@ -1,5 +1,0 @@
-import LoginForm from '@/components/login/LoginForm';
-
-export default function Login() {
-  return <LoginForm />;
-}
