--- resources/ts/pages/logout.tsx
@@ -1,15 +1,0 @@
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useAuth } from '@/hooks/useAuth';
-
-export default function Logout() {
-  const navigate = useNavigate();
-  const { logout } = useAuth();
-
-  useEffect(() => {
-    logout();
-    navigate('/login', { replace: true });
-  }, [logout, navigate]);
-
-  return null;
-}
