--- resources/ts/pages/worker/dashboard.tsx
@@ -1,7 +1,0 @@
-export default function Dashboard() {
-  return (
-    <div>
-      <h1>Worker Dashboard</h1>
-    </div>
-  );
-}
