+++ resources/ts/routes/AuthRoutes.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import UnauthenticatedRoute from '@/middlewares/UnauthenticatedRoute';
 
 import AuthLayout from '@/layouts/AuthLayout';
-import Login from '@/pages/login';
-import Logout from '@/pages/logout';
+import Login from '@/pages/Login';
+import Logout from '@/pages/Logout';
 
 import { RouteObject } from 'react-router-dom';
